verified correctness of counters and edge variables

+++ src/pc_emulator/include/elementary_datatypes.h
@@ -318,7 +318,7 @@ namespace pc_emulator {
 
     // Time data type. Contains seconds elapsed
     struct TIMEDataType {
-        s64 SecsElapsed;
+        double SecsElapsed;
 
 
         //! Add operator for two time data types
